test: align UX coverage with existing suites
